feat: requested changes, app template sync, deps update
