+++ build.rs
@@ -1,4 +1,3 @@
 fn main() {
-    gear_wasm_builder::build();
     gear_wasm_builder::build_with_metadata::<dao_io::DaoMetadata>();
 }
