+++ io/Cargo.toml
@@ -8,8 +8,8 @@ authors = ["Gear Technologies"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+gmeta = { git = "https://github.com/gear-tech/gear.git" }
 gstd = { git = "https://github.com/gear-tech/gear.git", branch = "stable" }
-gmeta = { git = "https://github.com/gear-tech/gear.git", branch = "master", features = ["codegen"] }
-scale-info = { version = "2.3.1", default-features = false }
-codec = { package = "parity-scale-codec", version = "3.2.1", default-features = false }
+scale-info = { version = "2", default-features = false }
+parity-scale-codec = { version = "3", default-features = false }
 hashbrown = "0.13.1"
