+++ state/Cargo.toml
@@ -8,7 +8,7 @@ authors = ["Gear Technologies"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-gstd = { git = "https://github.com/gear-tech/gear" }
+gstd = { git = "https://github.com/gear-tech/gear", branch = "stable" }
 gmeta = { git = "https://github.com/gear-tech/gear", features = ["codegen"] }
 dao-io = { path = "../io" }
